Move goToNextPaint/goToPrevPaint to redux actions
